x86_64 build fix.

+++ artpaint/controls/VisualColorControl.cpp
@@ -193,7 +193,7 @@ void VisualColorControl::Draw(BRect)
 void VisualColorControl::MessageReceived(BMessage *message)
 {
 	rgb_color *new_color;
-	int32 color_size;
+	ssize_t color_size;
 	switch (message->what) {
 		case B_PASTE:
 			if (message->FindData("RGBColor",B_RGB_COLOR_TYPE,(const void**)&new_color,&color_size) == B_OK) {
